Refine structured test list generator
Rework input to have each section be a separate test suite. Define the
type of the section as base, one off, or single parameter perturbation.
Right now machine and compiler are not perturbed.

Testing: seems to generate separate test into separate files with the
expected lists.

Todo: still need to consider if there is something special for
multi-parameter perturbation, or if it is just a one off.